fix: 修复异常和格式化

+++ postinstall.js
@@ -1,26 +1,28 @@
-'use strict';
+'use strict'
 
-const debug = require('debug')('elint-helpers:install');
-const path = require('path');
-const fs = require('fs-extra');
-const { linterConfigFile } = require('./config');
+const debug = require('debug')('elint-helpers:install')
+const path = require('path')
+const fs = require('fs-extra')
+const { linterConfigFile } = require('./config')
 
-debug('run install');
+debug('run install')
 
-const cwd = process.env.INIT_CWD;
-const root = cwd.split('/node_modules/')[0];
+const cwd = process.env.INIT_CWD
+const root = cwd.split('/node_modules/')[0]
 
-debug(`cwd: ${cwd}`);
-debug(`root: ${root}`);
-
-link(presetModulePath, keep);
+debug(`cwd: ${cwd}`)
+debug(`root: ${root}`)
 
 linterConfigFile.forEach(fileName => {
-  const from = path.join(cwd, fileName);
-  const to = path.join(root, fileName);
+  const from = path.join(cwd, fileName)
+  const to = path.join(root, fileName)
 
-  fs.moveSync(from, to, { overwrite: true });
+  if (!fs.pathExistsSync(from)) {
+    return
+  }
 
-  console.log(`  move: from "${from}"`);
-  console.log(`          to "${to}"`);
-});
+  fs.moveSync(from, to, { overwrite: true })
+
+  console.log(`  move: from "${from}"`)
+  console.log(`          to "${to}"`)
+})
